feat: Update Footer and Header components
The code changes in the Footer component update the text size to "text-xs" for better readability. In the Header component, the CSS class "hidden top-[5rem]" is added to improve the behavior of the navigation menu. These modifications aim to enhance the user experience and improve the overall appearance of the website.